requirements.txt
# Web Scraping
playwright>=1.40.0
aiohttp>=3.9.0

# LLM APIs
openai>=1.12.0
anthropic>=0.18.0

# Data Validation & Processing
pydantic>=2.6.0
pydantic-settings>=2.1.0

# Utilities
python-dotenv>=1.0.0
tenacity>=8.2.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Backend API
fastapi>=0.110.0
uvicorn>=0.27.0

# Optional but useful
httpx>=0.26.0
